requirements.txt
psycopg2-binary
flask
openai
python-dotenv
langchain
langchain-community
tiktoken
apscheduler
